Adding previous insert iterator and utils file
