+++ PartialModelSelectionMain.cpp
@@ -6,6 +6,7 @@
 #include <string>
 //Local Includes
 #include "PartialModelSelection.hpp"
+#include "ModelSelectionUtils.hpp"
 
 //Macros to display test info in gtest format. 
 #define GTEST_MINCOUT std::cerr << "[ MINIMIZE ] [INFO] "
